requirements.txt
absl-py==0.7.1
appnope==0.1.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
cloudpickle==1.2.1
colorama==0.3.9
configparser==4.0.2
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
easydict==1.9
efficientnet-pytorch==0.7.0
entrypoints==0.3
ffmpeg==1.4
future==0.17.1
gast==0.2.2
grpcio==1.22.0
gym==0.13.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.20
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
moviepy==1.0.0
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.3
notebook==6.0.1
numpy==1.16.4
numpy-quaternion==2019.7.23.15.26.49
opencv-python==4.4.0.44
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==4.1.0
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pybullet==2.5.1
pyglet==1.3.2
Pygments==2.4.2
pylint==2.3.1
pymesh==1.0.2
pyparsing==2.4.0
pyquaternion==0.9.5
pyrsistent==0.15.4
python-dateutil==2.8.0
PyWavefront==1.1.0
pyzmq==18.1.0
qtconsole==4.5.5
Quaternion==0.3.1
ratcave==0.8.0
requests==2.22.0
retrying==1.3.3
scipy==1.3.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
tensorboard==1.13.0
tensorboardX==1.4
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==1.6.0
torchvision==0.7.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
trimesh==2.38.40
typed-ast==1.4.0
urllib3==1.25.3
wavefront-reader==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.1
wordnet==0.0.1b2
wrapt==1.11.2
yattag==1.11.2
